backend-chatbot/app/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.3"
fastapi = {extras = ["standard"], version = "^0.115.5"}
openai = "0.27.0"
pdfquery = "^0.4.3"
pypdf2 = "^3.0.1"
pdfminer-six = "^20240706"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
